moved Groups interface inside Rules
