+++ Rules/Group/Changeable.ts
@@ -1,13 +1,13 @@
 import { isly } from "isly"
-import type { Rule } from "../index"
-import { type as RuleType } from "../type"
+import { Rule } from "../../Rule/index"
+// import { type as RuleType } from "../../Rule/type"
 
 export interface Changeable {
 	rules: Rule[]
 }
 export namespace Changeable {
 	export const type = isly.object<Changeable>({
-		rules: isly.array(RuleType),
+		rules: isly.array(Rule.type),
 	})
 	export const is = type.is
 	export const flaw = type.flaw
