+++ Rules/index.ts
@@ -1,19 +1,24 @@
 import { userwidgets } from "@userwidgets/model"
 import { isly } from "isly"
-import { Rule } from "../Rule/index"
 import { Changeable as RulesChangeable } from "./Changeable"
+import { Group as RulesGroup } from "./Group"
 
 export interface Rules extends Rules.Changeable {
-	common: Rule.Group
-	users: Record<userwidgets.Email, Rule.Group>
+	common: Rules.Group
+	users: Record<userwidgets.Email, Rules.Group>
 }
 export namespace Rules {
+	export type Group = RulesGroup
+	export const Group = RulesGroup
+	export namespace Group {
+		export type Changeable = RulesGroup.Changeable
+	}
 	export type Changeable = RulesChangeable
 	export const Changeable = RulesChangeable
 	export const type = isly.object<Rules>({
 		organization: userwidgets.Organization.Identifier.type,
-		common: Rule.Group.type,
-		users: isly.record<userwidgets.Email, Rule.Group>(userwidgets.Email.type, Rule.Group.type),
+		common: Group.type,
+		users: isly.record<userwidgets.Email, Group>(userwidgets.Email.type, Group.type),
 	})
 	export const is = type.is
 	export const flaw = type.flaw
